services/customer_support_agent/requirements.txt
transformers==4.52.4
torch==2.7.1
sentence-transformers==4.1.0
faiss-cpu==1.11.0
huggingface-hub==0.32.5
requests==2.32.4
openai==1.86.0
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.0.0